Stage A pre-launch hygiene: multi-root loader, splits, aug, bf16/DDP fixes
- accelerate_config: num_processes 1->2, MULTI_GPU for 2x A6000
- predictor.yaml: 4-root InterAct data, warmup 1000->200, loss weights
  rebalanced by init-loss scale, reachability_weight=0, prior_warmup 500
- train_predictor: ConcatDataset over 4 subsets, object-id split filter,
  SyncBatchNorm on object encoder, AdamW w/ no-decay groups, prior warmup
- HOIDataset: mirror / Y-rotation / pc-jitter augmentations behind
  AugmentConfig; build_object_split hash-based deterministic splitter
- losses: label_smoothing=0.1 on phase/support CE
- clip_utils: stop forcing .float() output (lets bf16 autocast flow)
- object_encoder: torch.cdist forced fp32 to avoid bf16 ball-query drift
- trainer: build_optimizer_with_decay_groups; step_fn gets global_step
  for prior warmup

+++ src/piano/models/object_encoder.py
@@ -138,8 +138,12 @@ class SetAbstractionLayer(nn.Module):
         M = new_xyz.shape[1]
         S = min(self.num_samples, N)
 
-        # Pairwise distances: (B, M, N)
-        dists = torch.cdist(new_xyz, xyz)
+        # Pairwise distances: (B, M, N). ``torch.cdist`` is numerically
+        # unstable under bf16 — ball-query neighbor sets near the
+        # radius boundary get randomized when two close points differ
+        # by less than bf16's precision. Force fp32 for the distance
+        # computation and cast back for indexing / gathering.
+        dists = torch.cdist(new_xyz.float(), xyz.float()).to(new_xyz.dtype)
 
         dists_sorted, idx_sorted = dists.sort(dim=-1)
         idx_sorted = idx_sorted[:, :, :S]  # (B, M, S)
